Make sure that the Last-Modified header is set correctly. (PR #738)
# Description

I noticed that we have caching issues on the web-console. It seems this is caused by Cloudflare dropping the `etag` header under some circumstances. I assume that several reverse proxies could the same thing. In this situation, the browser falls back to the `Last-Modified` header to determine whether the browser cache is still fresh. In the previous implementation this value was always to set to some fixed timestamp in 1985, because it's the modification timestamp of the file on disk. This PR fixes that issue by making sure that the `Last-Modified` header is set to the build date, included in the `version.json` file.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~