Added quickfix for the buildDate field in version.json file (PR #741)
# Description

On master the buildDate in the version.json file is formatted as `2025-08-03T01:41:52.686Z` on iso8 and iso7 it's formatted as `Mon Aug 04 2025 03:07:14 GMT+0200 (Central European Summer Time)`. This breaks the web-console on the dev build. This PR applies a quick fix on the iso8 and iso7 branches to fix the dev build. When Lukas is back (tomorrow), we should look into a permanent fix. I think it would be good if we align the format across the different iso versions and use the format from the master branch, because it's easier to parse and more standardized.

# Self Check:

- [ ] ~~Attached issue to pull request~~
- [x] Changelog entry
- [x] Type annotations are present
- [x] Code is clear and sufficiently documented
- [x] No (preventable) type errors
- [x] Sufficient test cases (reproduces the bug/tests the requested feature)
- [x] Correct, in line with design
- [ ] ~~End user documentation is included or an issue is created for end-user documentation~~